requirements.txt
Cython==0.22
PyYAML==3.10
argparse==1.2.1
h5py==2.5.0
numpy==1.9.2
oauth==1.0.1
pyconfig==2.2.1
pycurl==7.19.3
pyserial==2.6
requests==2.2.1
SQLAlchemy==1.0.4
python-swiftclient==2.4.0
mock==1.3.0
snakebite==2.8.2
